RavenDB-14509 cleanup
